+++ includes/classes/legacy/class-admin.php
@@ -62,7 +62,6 @@ class Admin extends Tour_Operator {
 
 		add_action( 'init', array( $this, 'init' ) );
 		add_action( 'admin_enqueue_scripts', array( $this, 'enqueue_admin_stylescripts' ), 999 );
-		add_action( 'cmb_save_custom', array( $this, 'post_relations' ), 3, 20 );
 	}
 
 	/**
@@ -72,6 +71,8 @@ class Admin extends Tour_Operator {
 	 */
 	public function init() {
 		if ( is_admin() ) {
+			add_action( 'cmb2_pre_save_field', array( $this, 'cpt_relations' ), 2, 20 );
+
 			$this->connections   = $this->create_post_connections();
 			$this->single_fields = apply_filters( 'lsx_to_search_fields', array() );
 			$this->taxonomies = apply_filters( 'lsx_to_taxonomies', $this->taxonomies );
@@ -135,100 +136,101 @@ class Admin extends Tour_Operator {
 	}
 
 	/**
-	 * Sets up the "post relations"
+	 * Fixes the CMB2 field relations
 	 *
-	 * @return    object
+	 * @param string $field_id
+	 * @param bool $status
+	 * @param string $action
+	 * @param object $field
+	 * @return void
 	 */
-	public function post_relations( $post_id, $field, $value ) {
-		if ( 'group' === $field['type'] && isset( $this->single_fields ) && array_key_exists( $field['id'], $this->single_fields ) ) {
-			$delete_counter = array();
+	public function cpt_relations( $field_id, $field ) {
+		
+		if ( in_array( $field_id, $this->connections ) ) {
+			$connected_id    = get_the_ID();
+			$previous_values = get_post_meta( $connected_id, $field_id, true );
+			$remote_key      = $this->reverse_key( $field_id );
 
-			foreach ( $this->single_fields[ $field['id'] ] as $fields_to_save ) {
-				$delete_counter[ $fields_to_save ] = 0;
+			if ( isset( $field->data_to_save[ $field_id ] ) ) {
+				$new_values = $field->data_to_save[ $field_id ];
+			} else {
+				$new_values = array();
 			}
 
-			//Loop through each group in case of repeatable fields
-			$relations          = false;
-			$previous_relations = false;
-
-			foreach ( $value as $group ) {
-				//loop through each of the fields in the group that need to be saved and grab their values.
-				foreach ( $this->single_fields[ $field['id'] ] as $fields_to_save ) {
-					//Check if its an empty group
-					if ( isset( $group[ $fields_to_save ] ) && ! empty( $group[ $fields_to_save ] ) ) {
-						if ( $delete_counter[ $fields_to_save ] < 1 ) {
-							//If this is a relation field, then we need to save the previous relations to remove any items if need be.
-							if ( in_array( $fields_to_save, $this->connections ) ) {
-								$previous_relations[ $fields_to_save ] = get_post_meta( $post_id, $fields_to_save, false );
-							}
-
-							delete_post_meta( $post_id, $fields_to_save );
-							$delete_counter[ $fields_to_save ] ++;
-						}
-
-						//Run through each group
-						foreach ( $group[ $fields_to_save ] as $field_value ) {
-							if ( null !== $field_value ) {
-								if ( 1 === $field_value ) {
-									$field_value = true;
-								}
-
-								add_post_meta( $post_id, $fields_to_save, $field_value );
-
-								//If its a related connection the save that
-								if ( in_array( $fields_to_save, $this->connections ) ) {
-									$relations[ $fields_to_save ][ $field_value ] = $field_value;
-								}
-							}
-						}
+			//if the new values are empty, then we need to remove the previous values.
+			if ( empty( $new_values ) ) {
+				if ( ! empty( $previous_values ) ) {
+					foreach ( $previous_values as $remote_id ) {
+						$this->remove_connected_id( $remote_id, $connected_id, $remote_key );
 					}
 				}
-			}//end of the repeatable group foreach
+			} else {
 
-			//If we have relations, loop through them and save the meta
-			if ( false !== $relations ) {
-				foreach ( $relations as $relation_key => $relation_values ) {
-					$temp_field = array(
-						'id' => $relation_key,
-					);
-
-					$this->save_related_post( $post_id, $temp_field, $relation_values, $previous_relations[ $relation_key ] );
+				if ( ! is_array( $previous_values ) ) {
+					$previous_values = [ $previous_values ];
 				}
-			}
-		} else {
-			if ( in_array( $field['id'], $this->connections ) ) {
-				$this->save_related_post( $post_id, $field, $value );
+
+				// Now determine if we added or removed any values.
+				$is_removing = array_diff( $previous_values, $new_values );
+				$is_adding   = array_diff( $new_values, $previous_values );
+
+				if ( ! empty( $is_removing ) ) {
+					foreach ( $is_removing as $remote_id ) {
+						$this->remove_connected_id( $remote_id, $connected_id, $remote_key );
+					}
+				}
+
+				if ( ! empty( $is_adding ) ) {
+					foreach ( $is_adding as $remote_id ) {
+						$this->add_connected_id( $remote_id, $connected_id, $remote_key );
+					}
+				}
 			}
 		}
 	}
 
 	/**
-	 * Save the reverse post relation.
+	 * Reverses the key for the remote_key slug.
 	 *
-	 * @return    null
+	 * @param [type] $meta_key
+	 * @return void
 	 */
-	public function save_related_post( $post_id, $field, $value, $previous_values = false ) {
-		$ids = explode( '_to_', $field['id'] );
-		$relation = $ids[1] . '_to_' . $ids[0];
+	public function reverse_key( $meta_key ) {
+		$ids = explode( '_to_', $meta_key );
+		return $ids[1] . '_to_' . $ids[0];
+	}
 
-		if ( in_array( $relation, $this->connections ) ) {
-			if ( false === $previous_values ) {
-				$previous_values = get_post_meta( $post_id, $field['id'], false );
-			}
+	/**
+	 * Remove the connected ID from the serialized array.
+	 *
+	 * @param int $remote_id
+	 * @param int $connected_id
+	 * @param string $meta_key
+	 * @return void
+	 */
+	public function remove_connected_id( $remote_id, $connected_id, $meta_key ) {
+		$prev = get_post_meta( $remote_id, $meta_key, true );
+		if ( ! empty( $prev ) ) {
+			$diff = array_diff( $prev, array( $connected_id ) );
+			update_post_meta( $remote_id, $meta_key, $diff, $prev );
+		}
+	}
 
-			if ( false !== $previous_values && ! empty( $previous_values ) ) {
-				foreach ( $previous_values as $tr ) {
-					delete_post_meta( $tr, $relation, $post_id );
-				}
+	public function add_connected_id( $remote_id, $connected_id, $meta_key ) {
+		$prev = get_post_meta( $remote_id, $meta_key, true );
+		// No Previous items detected.
+		if ( false === $prev || empty( $prev ) ) {
+			delete_post_meta( $remote_id, $meta_key );
+			$test = add_post_meta( $remote_id, $meta_key, array( $connected_id ), true );
+		} else {
+			if ( ! is_array( $prev ) ) {
+				$new = array( $prev );
+			} else {
+				$new = $prev;
 			}
-
-			if ( is_array( $value ) ) {
-				foreach ( $value as $v ) {
-					if ( '' !== $v && null !== $v && false !== $v ) {
-						add_post_meta( $v, $relation, $post_id );
-					}
-				}
-			}
+			$new[] = $connected_id;
+			$new   = array_unique( $new );
+			$updated = update_post_meta( $remote_id, $meta_key, $new, $prev );
 		}
 	}